DynamicTimeseries

Documentation for DynamicTimeseries.

DynamicTimeseries.AbstractDynamicDownsampler

DynamicTimeseries.CacheAccessor

DynamicTimeseries.Downsampler

Base.getindex

Base.getindex

DynamicTimeseries.downsamp_req

DynamicTimeseries.downsamp_req

DynamicTimeseries.reduce_downsample_caches

DynamicTimeseries.typemmap

SignalIndices.duration

SignalIndices.time_interval

DynamicTimeseries.AbstractDynamicDownsampler — Type

Must implement: downsampreq extrema fs baselength starttime

DynamicTimeseries.CacheAccessor — Method

Pulls parameters from inputs

DynamicTimeseries.Downsampler — Type

Must implement array interface binsize binbounds downsampreduce downsampreducecache

Base.getindex — Method

Method meant for when N > 1 in WindowedArray

Base.getindex — Method

Method meant for when N > 1 in WindowedArray

DynamicTimeseries.downsamp_req — Method

downsamp_req(ds, xb, xe, npt; windowfun)

Calculates the spectrogram for data in the range between xb and xe. Expands the selected range so the center of the first time bin is around xb, and the center of the last time bin is around xe.

DynamicTimeseries.downsamp_req — Method

downsamp_req(ds, xb, xe, npt; windowfun)

Calculates the spectrogram for data in the range between xb and xe. Expands the selected range so the center of the first time bin is around xb, and the center of the last time bin is around xe.

DynamicTimeseries.reduce_downsample_caches — Method

Recursively find downsampled values using cached arrays

DynamicTimeseries.typemmap — Method

Make a mmaped array of type A

SignalIndices.duration — Method

duration(a::AbstractDynamicDownsampler)

Returns the duration of a, in seconds.

SignalIndices.time_interval — Method

time_interval(a::AbstractDynamicDownsampler)

Returns the start and stop time of a, in seconds.